Refactor class names & added rendering interface to be passed to the Game and decouple it fro javaFX
