+++ PacMan/src/application/Main.java
@@ -1,11 +1,11 @@
 package application;
 	
 import javafx.application.Application;
-import rendering.RenderSystem;
+import rendering.RenderingSystem;
 
 public class Main {
 	
 	public static void main(String[] args) {
-		Application.launch(RenderSystem.class, args); 
+		Application.launch(RenderingSystem.class, args); 
 	}
 }
