+++ PacMan/src/rendering/IBoardRenderer.java
@@ -1,0 +1,10 @@
+package rendering;
+
+public interface IBoardRenderer {
+
+	void drawMaze();
+	
+	void spawnAnimatables();
+	
+	void spawnStaticEntities();
+}
